[ios] AST-66..69 stats page bundled features + consolidation
Consolidates duplicate `StatsView` / `FanficStatsView` into a new
shared `StatsUI` package; both tabs become ~10-line wrappers around
`SharedStatsView`. Adds four new sections powered by pure
`StatsAggregates` helpers:

- AST-66: Reading time card + weekly-minutes chart
  Uses LocalReadingSession.endedAt (already populated by reader
  dismiss handlers). Total time + avg session in the hero row,
  12-week minutes chart alongside the heatmap.

- AST-67: Top tags bar chart in the Breakdown section
  Parses tagsJSON (comics + fanfics) and freeformTags (fanfics),
  normalises lowercase, top 12 by frequency.

- AST-68: Heatmap day drill-down
  Heatmap cells are now tappable; opens a sheet listing sessions
  on that day with story title, chapter count, and duration.

- AST-69: Bookmark analytics section
  Summary counts, most-annotated stories (top 5), 12-week activity
  chart, and a horizontal highlights carousel of bookmarks with
  non-empty selectedText.

All data is local (SwiftData); no backend or model changes.
Package structure:
  ios/Astral/Packages/StatsUI/
    Sources/StatsUI/
      StatsData.swift              (pure aggregates, no UI)
      SharedStatsView.swift        (unified view + 5 sections)
      HeatmapDayDetailSheet.swift  (AST-68 drill-down sheet)

Registered in project.yml; xcodegen regenerated.

Fixes AST-66
Fixes AST-67
Fixes AST-68
Fixes AST-69
