+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Add a new entry whenever you merge a feature/fix PR into `development`. Keep ent
 
 ## 2026-04-22
 
+### iOS
+
+- **AST-66..69** — Stats page bundled features + consolidation. Extracted the duplicate-heavy `StatsView` (ComicFeature, 978 LOC) and `FanficStatsView` (FanficFeature, 696 LOC) into a new shared `StatsUI` Swift package; both tab wrappers are now ~10 lines delegating to `SharedStatsView`. Added four new sections fed by pure `StatsAggregates` helpers in `StatsData.swift`: (1) a **Time Read / Avg Session** odometer row plus a weekly-minutes chart alongside the activity heatmap (AST-66, sourced from `LocalReadingSession.endedAt` which was already being set on reader dismiss); (2) a **Top Tags** horizontal bar chart in the Breakdown section parsing `tagsJSON` from both comics and fanfics plus `freeformTags` on fanfics (AST-67); (3) **tappable heatmap cells** that open a day-detail sheet listing sessions with story title, chapter count, and session duration (AST-68); (4) a new **Bookmarks** analytics section with summary counts, top-annotated stories, 12-week bookmarking activity chart, and a horizontal highlights carousel showing `selectedText` previews (AST-69). All data is local-only; no backend or model changes. Fixes [AST-66](https://linear.app/nnetraganti/issue/AST-66), [AST-67](https://linear.app/nnetraganti/issue/AST-67), [AST-68](https://linear.app/nnetraganti/issue/AST-68), [AST-69](https://linear.app/nnetraganti/issue/AST-69).
+
 ### Backend
 
 - **AST-70..77** — Monitor dashboard control plane. Eight new developer-facing interactions on the OCI monitor: kill-switch panel for `MANGADEX_DISABLED` / `FFNET_NEW_SCRAPER_DISABLED` / `MANGADEX_AUTO_SWITCH_SOURCE` (Redis-backed overrides at `mon:flag:*`, consumed via new `app.core.runtime_flags.flag_enabled()` with 5s cache — no worker restart needed); one-click service restart (`/control/restart/{service}`, confirmation gate for stateful services); ARQ failed-job retry via fresh `enqueue_job`; SSE live log stream at `/metrics/logs/stream` (previously stubbed 501) with per-subscriber service/level/q filters fed off `LOG_DEQUE`; on-demand `pg_dump` via `docker exec postgres` writing to a new `monitor_state` named volume with polling + download endpoints; read-only SQL console against a new `astral_readonly` Postgres role (SELECT/WITH only, 10s statement_timeout, 500-row cap, history in localStorage); endpoint latency drill-down at `/metrics/endpoint` backed by per-endpoint buckets in `NGINX_ENDPOINT_CACHE` (top 20 per window); recent-deploy block reading `deploys.jsonl` appended by a new `backend/scripts/deploy-hook.sh`. All mutating endpoints gated on `X-Monitor-Auth` header / `MONITOR_CONTROL_TOKEN` env — 503 when unset, 401 on mismatch. Every action written to an append-only audit log surfaced in the dashboard. Docker socket widened to `rw`; new Alembic migration `a3b4c5d6e7f8` creates the `astral_readonly` role with SELECT grants + 10s timeout. UI augmented non-invasively via a new `/static/controls.js` companion that decorates the existing renderer via MutationObserver. Fixes [AST-70](https://linear.app/nnetraganti/issue/AST-70), [AST-71](https://linear.app/nnetraganti/issue/AST-71), [AST-72](https://linear.app/nnetraganti/issue/AST-72), [AST-73](https://linear.app/nnetraganti/issue/AST-73), [AST-74](https://linear.app/nnetraganti/issue/AST-74), [AST-75](https://linear.app/nnetraganti/issue/AST-75), [AST-76](https://linear.app/nnetraganti/issue/AST-76), [AST-77](https://linear.app/nnetraganti/issue/AST-77).
