+++ ios/Astral/Packages/ComicFeature/Package.swift
@@ -12,11 +12,12 @@ let package = Package(
         .package(path: "../Core"),
         .package(path: "../Networking"),
         .package(path: "../DesignSystem"),
+        .package(path: "../StatsUI"),
     ],
     targets: [
         .target(
             name: "ComicFeature",
-            dependencies: ["Core", "Networking", "DesignSystem"]
+            dependencies: ["Core", "Networking", "DesignSystem", "StatsUI"]
         ),
         .testTarget(
             name: "ComicFeatureTests",
